exp/pyproject.toml
[project]
name = "test-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fasta2a",
    "pydantic-ai>=1.67.0",
    "pydantic-ai-slim[mcp]>=1.80.0",
    "python-dotenv>=1.2.2",
]

[tool.uv.sources]
fasta2a = { git = "https://github.com/pydantic/fasta2a.git", rev = "b7b6f11d472745656d77c96d10e5656fc3be0e50" }

[tool.pyright]
venvPath = "."
venv = ".venv"
